Refactor Origin bridge modules
